+++ HW/hw08/Le7uk/Task08.1/models/__init__.py
@@ -1,0 +1,3 @@
+from .admin import create_admin
+from .user import create_user
+__all__ = ['create_admin', 'create_user']
