+++ HW/hw08/Le7uk/Task08.1/models/admin.py
@@ -1,0 +1,3 @@
+__all__ =["create_admin"]
+def create_admin():
+    pass
