+++ HW/hw08/Le7uk/Task08.1/models/user.py
@@ -1,0 +1,3 @@
+__all__ =["create_user"]
+def create_user():
+    pass
