+++ HW/hw08/Le7uk/Task08.1/utils/formaters.py
@@ -1,0 +1,3 @@
+__all__ =["format_string"]
+def format_string():
+    pass
