+++ HW/hw08/Le7uk/Task08.1/utils/logger.py
@@ -1,0 +1,3 @@
+__all__ =["log_in_file"]
+def log_in_file():
+    pass
