+++ HW/hw08/Le7uk/Task08.3/Calculations.py
@@ -1,0 +1,10 @@
+import math
+
+def calculate_area(radius):
+    return math.pi * pow(radius,2)
+
+def calculate_area_rectangle(length, width):
+    return length * width
+
+def calculate_area_triangle(base, height):
+    return 0.5 * base * height
