--- HW/hw08/Task08.1/models/__init__.py
@@ -1,3 +1,0 @@
-from .admin import create_admin
-from .user import create_user
-__all__ = ['create_admin', 'create_user']
