--- HW/hw08/Task08.1/models/admin.py
@@ -1,3 +1,0 @@
-__all__ =["create_admin"]
-def create_admin():
-    pass
