--- HW/hw08/Task08.1/models/user.py
@@ -1,3 +1,0 @@
-__all__ =["create_user"]
-def create_user():
-    pass
