--- HW/hw08/Task08.1/utils/__init__.py
@@ -1,3 +1,0 @@
-from .formaters import format_string
-from .logger import log_in_file
-__all__ = ['format_string', 'log_in_file']
