--- HW/hw08/Task08.1/utils/formaters.py
@@ -1,3 +1,0 @@
-__all__ =["format_string"]
-def format_string():
-    pass
