--- HW/hw08/Task08.1/utils/logger.py
@@ -1,3 +1,0 @@
-__all__ =["log_in_file"]
-def log_in_file():
-    pass
